backend/requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
Django==5.0
django-axes==8.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.48.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyotp==2.9.0
python-decouple==3.8
python-dotenv==1.2.1
rsa==4.9.1
sqlparse==0.5.5
